+++ km/com/hifiremote/jp1/GI2.java
@@ -10,13 +10,13 @@ public class GI2
 
   public void in( Value[] parms, Hex hexData, DeviceParameter[] devParms, int onlyIndex )
   {
-    byte[] hex = hexData.getData();
+    int[] hex = hexData.getData();
     int dev = (( Integer )devParms[0].getValueOrDefault()).intValue();
     int obc = (( Integer )parms[ 0 ].getUserValue()).intValue();
     int cmd = (dev<<6)+obc;
-    hex[0] = reverse( (byte) (63-obc) );
+    hex[0] = reverse( 63 - obc );
     int check = (cmd<<6) ^ (cmd<<3) ^ (cmd<<2) ^ cmd ^ (cmd>>2) ^ (cmd>>3) ^ (cmd>>4);
-    hex[1] = reverse( (byte)( 63 - dev - (check&56) - ((check>>4)&4) ) );
+    hex[1] = reverse( 63 - dev - (check&56) - ((check>>4)&4) );
   }
 
   public void out( Hex hex, Value[] parms, DeviceParameter[] devParms )
